app_server/requirements.txt
openai==1.6.0
opencv-python-headless==4.8.1.78
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
streamlit==1.29.0
streamlit-lottie==0.0.5
